+++ src/PatchKit.Logging/IMessageWriter.cs
@@ -1,7 +1,0 @@
-﻿namespace PatchKit.Logging
-{
-    public interface IMessageWriter
-    {
-        void Write(Message message, MessageContext messageContext);
-    }
-}